Wrapper
Create new components for the wrapper (header, footer & Content
