+++ .eslintcache
@@ -1,1 +1,1 @@
-[{"/Users/cathydolle/Documents/GitHub/portfolio2021/src/reportWebVitals.js":"1","/Users/cathydolle/Documents/GitHub/portfolio2021/src/Routes.js":"2","/Users/cathydolle/Documents/GitHub/portfolio2021/src/pages/Menu.jsx":"3","/Users/cathydolle/Documents/GitHub/portfolio2021/src/components/MenuIcon.jsx":"4","/Users/cathydolle/Documents/GitHub/portfolio2021/src/App.js":"5","/Users/cathydolle/Documents/GitHub/portfolio2021/src/components/Background.jsx":"6","/Users/cathydolle/Documents/GitHub/portfolio2021/src/pages/HomePage.jsx":"7","/Users/cathydolle/Documents/GitHub/portfolio2021/src/components/Logo.jsx":"8"},{"size":362,"mtime":1606582745416,"results":"9","hashOfConfig":"10"},{"size":278,"mtime":1606582745406,"results":"11","hashOfConfig":"10"},{"size":498,"mtime":1606582745416,"results":"12","hashOfConfig":"10"},{"size":772,"mtime":1606582745415,"results":"13","hashOfConfig":"10"},{"size":952,"mtime":1606583883604,"results":"14","hashOfConfig":"10"},{"size":670,"mtime":1606583083925,"results":"15","hashOfConfig":"10"},{"size":2914,"mtime":1606584460585,"results":"16","hashOfConfig":"10"},{"size":650,"mtime":1606584584274,"results":"17","hashOfConfig":"10"},{"filePath":"18","messages":"19","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"1pds9o0",{"filePath":"20","messages":"21","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"22","messages":"23","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"24","messages":"25","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"26","messages":"27","errorCount":0,"warningCount":1,"fixableErrorCount":0,"fixableWarningCount":0,"source":null},{"filePath":"28","messages":"29","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"30","messages":"31","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"32","messages":"33","errorCount":0,"warningCount":1,"fixableErrorCount":0,"fixableWarningCount":0,"source":null},"/Users/cathydolle/Documents/GitHub/portfolio2021/src/reportWebVitals.js",[],"/Users/cathydolle/Documents/GitHub/portfolio2021/src/Routes.js",[],"/Users/cathydolle/Documents/GitHub/portfolio2021/src/pages/Menu.jsx",[],"/Users/cathydolle/Documents/GitHub/portfolio2021/src/components/MenuIcon.jsx",[],"/Users/cathydolle/Documents/GitHub/portfolio2021/src/App.js",["34"],"/Users/cathydolle/Documents/GitHub/portfolio2021/src/components/Background.jsx",[],"/Users/cathydolle/Documents/GitHub/portfolio2021/src/pages/HomePage.jsx",[],"/Users/cathydolle/Documents/GitHub/portfolio2021/src/components/Logo.jsx",["35"],{"ruleId":"36","severity":1,"message":"37","line":18,"column":16,"nodeType":"38","messageId":"39","endLine":18,"endColumn":23},{"ruleId":"40","severity":1,"message":"41","line":4,"column":1,"nodeType":"42","endLine":10,"endColumn":3},"no-unused-vars","'setDark' is assigned a value but never used.","Identifier","unusedVar","import/no-anonymous-default-export","Assign arrow function to a variable before exporting as module default","ExportDefaultDeclaration"]
+[{"/Users/cathydolle/Documents/GitHub/portfolio2021/src/reportWebVitals.js":"1","/Users/cathydolle/Documents/GitHub/portfolio2021/src/Routes.js":"2","/Users/cathydolle/Documents/GitHub/portfolio2021/src/pages/Menu.jsx":"3","/Users/cathydolle/Documents/GitHub/portfolio2021/src/App.js":"4","/Users/cathydolle/Documents/GitHub/portfolio2021/src/components/Background.jsx":"5","/Users/cathydolle/Documents/GitHub/portfolio2021/src/pages/HomePage.jsx":"6","/Users/cathydolle/Documents/GitHub/portfolio2021/src/components/wrapper/Wrapper.jsx":"7","/Users/cathydolle/Documents/GitHub/portfolio2021/src/components/icons/MenuIcon.jsx":"8","/Users/cathydolle/Documents/GitHub/portfolio2021/src/components/wrapper/Footer.jsx":"9","/Users/cathydolle/Documents/GitHub/portfolio2021/src/components/wrapper/Content.jsx":"10"},{"size":362,"mtime":1606582745416,"results":"11","hashOfConfig":"12"},{"size":278,"mtime":1606582745406,"results":"13","hashOfConfig":"12"},{"size":498,"mtime":1606582745416,"results":"14","hashOfConfig":"12"},{"size":952,"mtime":1606583883604,"results":"15","hashOfConfig":"12"},{"size":670,"mtime":1606583083925,"results":"16","hashOfConfig":"12"},{"size":2920,"mtime":1606584844754,"results":"17","hashOfConfig":"12"},{"size":710,"mtime":1606586038680,"results":"18","hashOfConfig":"12"},{"size":772,"mtime":1606584861588,"results":"19","hashOfConfig":"12"},{"size":275,"mtime":1606585775353,"results":"20","hashOfConfig":"12"},{"size":365,"mtime":1606585970730,"results":"21","hashOfConfig":"12"},{"filePath":"22","messages":"23","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"1pds9o0",{"filePath":"24","messages":"25","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"26","messages":"27","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"28","messages":"29","errorCount":0,"warningCount":1,"fixableErrorCount":0,"fixableWarningCount":0,"source":null},{"filePath":"30","messages":"31","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"32","messages":"33","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"34","messages":"35","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"36","messages":"37","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"38","messages":"39","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"40","messages":"41","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"/Users/cathydolle/Documents/GitHub/portfolio2021/src/reportWebVitals.js",[],"/Users/cathydolle/Documents/GitHub/portfolio2021/src/Routes.js",[],"/Users/cathydolle/Documents/GitHub/portfolio2021/src/pages/Menu.jsx",[],"/Users/cathydolle/Documents/GitHub/portfolio2021/src/App.js",["42"],"/Users/cathydolle/Documents/GitHub/portfolio2021/src/components/Background.jsx",[],"/Users/cathydolle/Documents/GitHub/portfolio2021/src/pages/HomePage.jsx",[],"/Users/cathydolle/Documents/GitHub/portfolio2021/src/components/wrapper/Wrapper.jsx",[],"/Users/cathydolle/Documents/GitHub/portfolio2021/src/components/icons/MenuIcon.jsx",[],"/Users/cathydolle/Documents/GitHub/portfolio2021/src/components/wrapper/Footer.jsx",[],"/Users/cathydolle/Documents/GitHub/portfolio2021/src/components/wrapper/Content.jsx",[],{"ruleId":"43","severity":1,"message":"44","line":18,"column":16,"nodeType":"45","messageId":"46","endLine":18,"endColumn":23},"no-unused-vars","'setDark' is assigned a value but never used.","Identifier","unusedVar"]
