+++ src/components/wrapper/Footer.jsx
@@ -1,0 +1,17 @@
+import React from 'react';
+import styled from '@emotion/styled';
+
+const FooterContent = styled.footer`
+  width: 100%;
+  height: 15%;
+`;
+
+function Footer() {
+    return (
+        <FooterContent>
+            <p>Test</p>
+        </FooterContent>
+    );
+}
+
+export default Footer;
